Added smoke engine smoke sound;

+++ Assets/Scripts/OtherStaff/AudioPooler/AudioPoolItem.cs
@@ -8,7 +8,7 @@ public class AudioPoolItem : MonoBehaviour
     public AudioSource audioSource;
     public TransformPositionLinker positionLinker;
     public Tween waitTween;
-    public float priority;
+    public int priority;
     public int ID;
 
     #region MonoBehaviour
